refactor and docker configs added

+++ app/controllers/userController.js
@@ -5,6 +5,17 @@ const { Op } = require("sequelize");
 const User = require("../models/user.js");
 
 const userController = (() => {
+    const wrapErrorsFunction = (foo) => {
+        return async (req, res) => {
+            try{
+                await foo(req, res);
+            }catch(e){
+                console.log(`SERVER ERROR ${e}`);
+                res.status(500).send({ message: "Internal server error" });
+            }
+        }
+    }
+
 	const userPropCheck = (user, res) => {
 		let result = !['userName', 'email', 'password'].every(el => el in user);
 		if (result) res.status(400).send({ message: "Invalid input data for route" });
@@ -21,53 +32,54 @@ const userController = (() => {
 			user.password = hashPassword;
 			try {
 				await User.create(user);
-				res.status(201).send(
-					{
-						userName: user.userName,
-					}
-				);
+                res.status(201).send({userName: user.userName,});
 			} catch (e) {
-				res.status(201).send(
+				res.status(409).send(
 					{
 						userName: `User with username: '${user.userName}' already exist!`,
 					}
 				);
 			}
-		} catch {
+		} catch(e) {
+            console.log(`SERVER ERROR ${e}`);
 			res.status(500).send({ message: "Internal server error" });
 		}
-	};
+	}
+
+    const findUserInDB = async (user, res) =>{
+        let dbUser = await User.findOne({
+            where: {
+                userName: user.userName,
+                [Op.or]: {
+                    email: user.email,
+                },
+            },
+        });
+
+        return dbUser || res.status(404).send({ message: `Querried user, ${user.userName}, does not exist!` });
+    }
 
 	const login = async (req, res) => {
 		let inUser = req.body;
 
-		if (userPropCheck(inUser, res)) return;
+		if (!(('userName' in inUser || 'email' in inUser) && 'password' in inUser)){ 
+            res.status(400).send({ message: "Invalid input data for route" });
+            return;
+        }
 
 		try {
-			let dbUser = await User.findOne({
-				where: {
-					userName: inUser.userName,
-					[Op.or]: {
-						email: inUser.email,
-					},
-				},
-			});
-
-			if (dbUser == null) {
-				res.status(404).send({ message: "Querried user does not exist!" });
-				return;
-			}
+           let dbUser = findUserInDB(inUser, res);
+			if (dbUser) return;
 
 			if (await bcrypt.compare(inUser.password, dbUser.password)) {
 				inUser.id = dbUser.id;
 				inUser.password = dbUser.password;
 				let token = tokenAuth.create(inUser);
 				res.status(200).send({ token: token, user: dbUser });
-			} else {
-				res.status(409).send({ message: "Unsuccessful login" });
-			}
+			} else res.status(409).send({ message: "Unsuccessful login" });
 		} catch (e) {
-			res.status(400).send({ message: "Invalid login data!" });
+            console.log(`SERVER ERROR ${e}`);
+			res.status(500).send({ message: "Internal server error" });
 		}
 	};
 
@@ -79,14 +91,9 @@ const userController = (() => {
 		if (userPropCheck(inUser, res)) return;
 
 		try {
-			let dbUser = await User.findOne({
-				where: {
-					userName: inUser.userName,
-					[Op.or]: {
-						email: inUser.email,
-					},
-				},
-			});
+            let dbUser = findUserInDB(inUser, res);
+			if (dbUser) return;
+
 			if (await bcrypt.compare(inUser.password, dbUser.password)) {
 				await User.destroy({
 					where: {
@@ -109,16 +116,24 @@ const userController = (() => {
 
 
 	const getUser = async (req, res) => {
-		let { userName } = req.body;
+		let inUser = req.body;
+        let isGood = false;
+        ['email', 'userName'].forEach(el => {
+            if(!(el in inUser)) inUser[el] = '';
+            else isGood = true;
+        });
+        if(!isGood){
+			res.status(400).send({ message: "Invalid input data for route" });
+            return;
+        }
+
 		try {
-			let dbUser = await User.findOne({
-				where: {
-					userName: userName
-				}
-			});
+			let dbUser = findUserInDB(inUser, res);
+			if (dbUser) return;
 
 			res.status(200).send(dbUser);
 		} catch (e) {
+            console.log(`SERVER ERROR ${e}`);
 			res.status(500).send({ message: "Internal server error" });
 		}
 	};
@@ -130,11 +145,14 @@ const userController = (() => {
 			});
 			res.status(200).send({ allUsers: allUser.map(el => el.userName) });
 		} catch (e) {
+            console.log(`SERVER ERROR ${e}`);
 			res.status(500).send({ message: "Internal server error" });
 		}
 	}
 
-	return { register, login, deleteUser, getUser, getAllUsers };
+    const returnObj = {};
+	[register, login, deleteUser, getUser, getAllUsers].forEach(foo => returnObj[foo.name] = foo);//wrapErrorsFunction(foo));
+    return returnObj;
 })();
 
 module.exports = userController;
